refactor(realtime,desktop,mobile): one subscriber component, one mobile hosts hook
The realtime Worker imports verifyJWT from @superset/shared instead of its
own copy. The desktop's useRealtimeNudges hook existed only for the
RealtimeNudges component, so the component is the whole of it now. Mobile's
useOrgHosts returns the presence-merged rows and the raw query together, so
screens stop importing two hooks for one list.
