--- apps/desktop/src/renderer/hooks/useRealtimeNudges/index.ts
@@ -1,1 +1,0 @@
-export { useRealtimeNudges } from "./useRealtimeNudges";
