+++ apps/realtime/package.json
@@ -12,7 +12,6 @@
 		"@sentry/cloudflare": "10.68.0",
 		"@superset/shared": "workspace:*",
 		"hono": "4.13.5",
-		"jose": "6.2.2",
 		"partyserver": "0.5.9"
 	},
 	"devDependencies": {
