+++ packages/shared/package.json
@@ -252,12 +252,13 @@
 			"types": "./src/host-routing.ts",
 			"default": "./src/host-routing.ts"
 		},
-		"./verify-jwt": {
-			"types": "./src/verify-jwt.ts",
-			"default": "./src/verify-jwt.ts"
 		"./realtime": {
 			"types": "./src/realtime.ts",
 			"default": "./src/realtime.ts"
+		},
+		"./verify-jwt": {
+			"types": "./src/verify-jwt.ts",
+			"default": "./src/verify-jwt.ts"
 		},
 		"./install-update-lock": {
 			"types": "./src/install-update-lock.ts",
